Move button into input folder
